[Add] gateway presences

+++ src/sharding/ShardingManager.ts
@@ -10,6 +10,7 @@ export declare interface ShardingManager {
 }
 export class ShardingManager extends EventEmitter {
     shards: Map<number, ChildProcess> = new Map();
+    private file: string = null;
 
     constructor(
         public token: string,
@@ -34,32 +35,37 @@ export class ShardingManager extends EventEmitter {
         }
     }
 
+    config(config: { file: string }) {
+        this.file = config.file;
+    }
+
     async spawn() {
+        if(!this.file) throw new Error(`No file provided to spawn shards from\n${'HELP'.green} -> Use ShardingManager.config({ file: "path/to/file" }) to set the file path\n`);
         let shardCount = await this.getShardCount();
         for (let i = 0; i < shardCount; i++) {
             let cp = fork('', {
                 stdio: 'inherit',
                 execArgv: [
-                    `${__dirname}/Shard.js`,
+                    `${this.file}`,
                     `${i}`,
                     `${shardCount}`,
-                    `${this.token}`,
                     `${this.intents}`
                 ]
             });
             cp.on('error', () => {
                 cp.kill();
+                // TODO respawn shard
             });
-            cp.on('message', (msg: any /* do this type later */) => {
-                switch (msg.type) {
-                    case 'ready':
-                        this.emit('shardReady', msg.data);
-                        break;
+            // cp.on('message', (msg: any /* do this type later */) => {
+            //     switch (msg.type) {
+            //         case 'ready':
+            //             this.emit('shardReady', msg.data);
+            //             break;
                 
-                    default:
-                        break;
-                }
-            })
+            //         default:
+            //             break;
+            //     }
+            // })
 
             this.shards.set(i, cp);
             await sleep(5000);
